Preparing 1.8.1 (#370)
* Update Whisper.net dependencies to version 1.8.0

Updated all example projects to use Whisper.net and relevant runtime packages version 1.8.0. Adjusted code to use the new `WhisperGgmlDownloader.Default` API for model downloads and manifest file retrieval. Incremented the library version to 1.8.1-preview1 in the main project.

* Fixed windows AVX build path

* Updated readme to 1.8.1

* Improved tests to catch https://github.com/sandrohanea/whisper.net/issues/371 type of issues in the future

* Introduced vulkan, openvino and coreml to dotnet jobs as dependencies

* Added noavx CI on dotnet-noavx.yml

* try again

* Add CpuNoAvx parameter to Dependency Checker steps

Updated the workflow to run the Dependency Checker with the `CpuNoAvx` parameter in the dotnet-noavx.yml

* Add condition for NoAVX tests in project imports

This update introduces a new conditional import for handling NoAVX test scenarios in the project configuration. It ensures the correct runtime targets are used based on the `USE_WHISPER_NOAVX_TESTS` flag. Additionally, the `TargetFrameworks` property was corrected to `TargetFramework`.

* Increased timeout and renamed avx artifacts for test results

---------

Co-authored-by: Sandro Hanea <me@sandro.rocks>